Update for new gcc-4 warning options.
Added several new options for the gcc-4 generation, and updated the source
accordingly. This includes some code restructuring to eliminate unreachable
code, elimination of some gotos, elimination of unused return values, and
some additional casting.

+++ source/compiler/aslanalyze.c
@@ -667,8 +667,8 @@ AnCheckForReservedName (
     {
         /* The next two characters must be hex digits */
 
-        if ((isxdigit (Name[2])) &&
-            (isxdigit (Name[3])))
+        if ((isxdigit ((int) Name[2])) &&
+            (isxdigit ((int) Name[3])))
         {
             return (ACPI_EVENT_RESERVED_NAME);
         }
@@ -1236,7 +1236,7 @@ AnMethodAnalysisWalkBegin (
                      */
                     for (i = 0; Next->Asl.Value.String[i]; i++)
                     {
-                        if (!isalnum (Next->Asl.Value.String[i]))
+                        if (!isalnum ((int) Next->Asl.Value.String[i]))
                         {
                             AslError (ASL_ERROR, ASL_MSG_ALPHANUMERIC_STRING,
                                 Next, Next->Asl.Value.String);
